requirements.txt
streamlit
pandas
numpy
thefuzz
python-levenshtein
openpyxl
google-generativeai
scikit-learn
